add mapper trait
