+++ libzoea/src/syscall.rs
@@ -6,7 +6,8 @@ use crate::SysCallNo;
 use core::arch::asm;
 pub use shared::cap_type::CapabilityType;
 
-pub type SysCallRes = Result<usize, (ErrKind, u16)>;
+pub type SysCallRes = Result<usize, SysCallFailed>;
+pub type SysCallFailed = (ErrKind, u16)
 
 #[allow(clippy::too_many_arguments)]
 unsafe fn syscall(
